#37 refactor: 멤버 관련 테스트 코드 복구
